Phase5/requirements.txt
psycopg2-binary==2.9.7
tkinter
Pillow==10.0.0
ttkthemes==3.2.2
